[FIX] unstable behaviour on invoice cancel thru refund
This PR try fix an unstable behaviour when you're canceling an invoice through a refund. Some time the refund tackled first, some time it's the invoice. So it lead to an refund already marked as paid and though was filtered... The process wasn't detecting that there was a canceling process and was processing the invoice matching as a paid invoice.